reorganize and rename files
