+++ src/controllers/commands/addCommand.js
@@ -1,5 +1,5 @@
-const message = require("../../functions/message");
-const Command = require('../../models/Command');
+const message = require("../../util/message");
+const Command = require('../../definitions/Command');
 
 const go = (msg, rest) => {
     message.send(msg.channel, "I can't do that, sorry.");
